requirements.txt
flask
python-dotenv
requests
google-generativeai
pandas
openpyxl
pydub
twilio
gunicorn
gspread
oauth2client
